backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==3.2.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
websockets==12.0
requests==2.31.0
python-dotenv==1.0.0
email-validator